requirements.txt
#
#                       IsThicc-bot Requirements.txt | 2020-2021 (c) IsThicc
#
# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
#
#
jishaku
git+https://github.com/Rapptz/discord.py.git # Newer discord.py for reply support
aiomysql
discord-py-slash-command
#
#
# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
#
#